+++ packages/generator/src/service/csn.ts
@@ -1,4 +1,4 @@
-import fs = require('fs');
+import fs from 'fs';
 import edm2csn = require('@sap/edm-converters/lib/edmToCsn/lib/main');
 import { VdmServiceMetadata } from '../vdm-types';