+++ test-packages/type-tests/test/v4/batch.spec.ts
@@ -1,7 +1,7 @@
 import {
   batch,
   testService
-} from '@sap-cloud-sdk/test-services/v4/test-service';
+} from '@sap-cloud-sdk/test-services-odata-v4/test-service';
 import { ReadResponse } from '@sap-cloud-sdk/odata-v4/internal';
 import { customTestDeSerializers } from '../../../../test-resources/test/test-util';